requirements.txt
cachetools==5.2.0
google-auth==2.11.0
google-auth-oauthlib==0.5.2
gspread==5.5.0
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.1
rsa==4.9
